requirements.txt
flask==3.0.0
requests==2.31.0
pytest==7.4.0
flasgger
allure-pytest==2.13.2
